refactor: reorganize test utilities and improve documentation
