+++ pkg/config-manager/kind/manager_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 
 	configmanager "github.com/devantler-tech/ksail-go/pkg/config-manager"
-	"github.com/devantler-tech/ksail-go/pkg/config-manager/helpers/testutils"
 	"github.com/devantler-tech/ksail-go/pkg/config-manager/kind"
+	"github.com/devantler-tech/ksail-go/pkg/config-manager/testutils"
 	"github.com/stretchr/testify/assert"
 	v1alpha4 "sigs.k8s.io/kind/pkg/apis/config/v1alpha4"
 )
