+++ pkg/config-manager/testutils/helpers.go
